+++ src/common/posts/components/CreatePost.tsx
@@ -16,13 +16,13 @@ import {
   Textarea,
 } from "@chakra-ui/react";
 
-import useUserStore from "../../../../store/user.zustand";
+import useUserStore from "../../../store/user.zustand";
 import { CloseIcon, PlusSquareIcon } from "@chakra-ui/icons";
-import { ThemeColor } from "../../../../common/styles/theme.style";
+import { ThemeColor } from "../../styles/theme.style";
 import { Image } from "@chakra-ui/image";
 import styled from "@emotion/styled";
-import postApi from "../../../../api/postApi";
-import { CreatePostDto } from "../../../../api/dtos/post.dto";
+import postApi from "../../../api/postApi";
+import { CreatePostDto } from "../../../api/dtos/post.dto";
 
 const CreatePost = () => {
   //call user_id from zustand
