+++ src/common/posts/components/commentList.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box } from "@chakra-ui/layout";
 import Comment from "./comment";
-import { QueryCommentDto } from "../../../../api/dtos/comment.dto";
+import { QueryCommentDto } from "../../../api/dtos/comment.dto";
 
 interface CommentListProps {
   comments: QueryCommentDto[];
