+++ src/pages/profile/Profile.tsx
@@ -1,18 +1,16 @@
 import React, { Suspense, useState } from "react";
 import { useParams } from "react-router";
 
-import BasicPageLayout from "../../../common/components/layouts/BasicPageLayout";
+import BasicPageLayout from "../../common/components/layouts/BasicPageLayout";
 
 import ProfileCard from "./components/ProfileCard";
-import Reps from "../posts/Posts";
-import repsApi from "../../../api/postApi";
-import userApi from "../../../api/userApi";
+import repsApi from "../../api/postApi";
+import userApi from "../../api/userApi";
 import { QueryErrorResetBoundary, useQuery } from "@tanstack/react-query";
 import { ErrorBoundary } from "react-error-boundary";
-import ErrorPage from "../../../common/components/ErrorPage";
-import BlueSpinner from "../../../common/components/spinners/BlueSpinner";
-import { QueryPostDto } from "../../../api/dtos/post.dto";
-import Posts from "../posts/Posts";
+import ErrorPage from "../../common/components/ErrorPage";
+import BlueSpinner from "../../common/components/spinners/BlueSpinner";
+import Posts from "../../common/posts/Posts";
 
 const Profile = () => {
   const username = useParams().username;
